# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SDN/AAC Internal Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /financial-movements:
    get:
      operationId: listFinancialMovements
      tags: [financialMovements]
      parameters:
        - name: seasonId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Financial movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinancialMovement"
    post:
      operationId: createFinancialMovement
      tags: [financialMovements]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialMovementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialMovement"
  /financial-movements/{id}:
    get:
      operationId: getFinancialMovement
      tags: [financialMovements]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Financial movement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialMovement"
    patch:
      operationId: updateFinancialMovement
      tags: [financialMovements]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancialMovementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialMovement"
    delete:
      operationId: deleteFinancialMovement
      tags: [financialMovements]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /financial-summary:
    get:
      operationId: getFinancialSummary
      tags: [financialMovements]
      parameters:
        - name: seasonId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Financial summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialSummary"

  # ── Quota Plans ───────────────────────────────────────────────────────────────
  /quota-plans:
    get:
      operationId: listQuotaPlans
      tags: [quotaPlans]
      parameters:
        - name: seasonId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Quota plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuotaPlan"
    post:
      operationId: createQuotaPlan
      tags: [quotaPlans]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotaPlanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotaPlan"
  /quota-plans/{id}:
    patch:
      operationId: updateQuotaPlan
      tags: [quotaPlans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotaPlanUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotaPlan"
    delete:
      operationId: deleteQuotaPlan
      tags: [quotaPlans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Quotas ────────────────────────────────────────────────────────────────────
  /quotas:
    get:
      operationId: listQuotas
      tags: [quotas]
      parameters:
        - name: seasonId
          in: query
          schema:
            type: integer
        - name: athleteId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Quotas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quota"
  /quotas/generate:
    post:
      operationId: generateQuotas
      tags: [quotas]
      summary: Generate quotas for all athletes based on the quota plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateQuotasInput"
      responses:
        "200":
          description: Generated quotas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quota"
  /quotas/{id}:
    get:
      operationId: getQuota
      tags: [quotas]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quota
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quota"

  # ── Payments ──────────────────────────────────────────────────────────────────
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      parameters:
        - name: quotaId
          in: query
          schema:
            type: integer
        - name: athleteId
          in: query
          schema:
            type: integer
        - name: seasonId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    delete:
      operationId: deletePayment
      tags: [payments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quotas-overdue:
    get:
      operationId: listOverdueQuotas
      tags: [quotas]
      summary: List athletes with overdue or pending quotas
      parameters:
        - name: seasonId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Overdue quota summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AthleteQuotaSummary"

  # ── Documents ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    FinancialMovement:
      type: object
      required: [id, type, category, amount, date, description]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [receita, despesa]
        category:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date
        description:
          type: string
        documentUrl:
          type: ["string", "null"]
        seasonId:
          type: ["integer", "null"]
        seasonName:
          type: ["string", "null"]
        relatedQuotaId:
          type: ["integer", "null"]
    FinancialMovementInput:
      type: object
      required: [type, category, amount, date, description]
      properties:
        type:
          type: string
          enum: [receita, despesa]
        category:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date
        description:
          type: string
        documentUrl:
          type: ["string", "null"]
        seasonId:
          type: ["integer", "null"]
    FinancialMovementUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [receita, despesa]
        category:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date
        description:
          type: string
        documentUrl:
          type: ["string", "null"]
        seasonId:
          type: ["integer", "null"]
    FinancialSummary:
      type: object
      required: [totalRevenue, totalExpenses, balance]
      properties:
        totalRevenue:
          type: number
        totalExpenses:
          type: number
        balance:
          type: number
        byCategory:
          type: array
          items:
            type: object
            required: [category, type, total]
            properties:
              category:
                type: string
              type:
                type: string
              total:
                type: number

    # ── Quota Plans ─────────────────────────────────────────────────────────────
    QuotaPlan:
      type: object
      required: [id, seasonId, category, amount, periodicity]
      properties:
        id:
          type: integer
        seasonId:
          type: integer
        seasonName:
          type: ["string", "null"]
        category:
          type: string
        amount:
          type: number
        periodicity:
          type: string
          enum: [anual, mensal, trimestral]
        dueDay:
          type: ["integer", "null"]
    QuotaPlanInput:
      type: object
      required: [seasonId, category, amount, periodicity]
      properties:
        seasonId:
          type: integer
        category:
          type: string
        amount:
          type: number
        periodicity:
          type: string
          enum: [anual, mensal, trimestral]
        dueDay:
          type: ["integer", "null"]
    QuotaPlanUpdate:
      type: object
      properties:
        amount:
          type: number
        periodicity:
          type: string
          enum: [anual, mensal, trimestral]
        dueDay:
          type: ["integer", "null"]

    # ── Quotas ──────────────────────────────────────────────────────────────────
    Quota:
      type: object
      required: [id, athleteId, seasonId, amountDue, amountPaid, status]
      properties:
        id:
          type: integer
        athleteId:
          type: integer
        athleteName:
          type: ["string", "null"]
        seasonId:
          type: integer
        seasonName:
          type: ["string", "null"]
        period:
          type: ["string", "null"]
        amountDue:
          type: number
        amountPaid:
          type: number
        amountOwed:
          type: number
        status:
          type: string
          enum: [pago, parcial, pendente, em_atraso]
        dueDate:
          type: ["string", "null"]
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
    GenerateQuotasInput:
      type: object
      required: [seasonId, quotaPlanId]
      properties:
        seasonId:
          type: integer
        quotaPlanId:
          type: integer
    AthleteQuotaSummary:
      type: object
      required: [athleteId, athleteName, totalDue, totalPaid, totalOwed]
      properties:
        athleteId:
          type: integer
        athleteName:
          type: string
        totalDue:
          type: number
        totalPaid:
          type: number
        totalOwed:
          type: number
        overdueCount:
          type: integer
        quotas:
          type: array
          items:
            $ref: "#/components/schemas/Quota"

    # ── Payments ────────────────────────────────────────────────────────────────
    Payment:
      type: object
      required: [id, quotaId, amount, date]
      properties:
        id:
          type: integer
        quotaId:
          type: integer
        amount:
          type: number
        date:
          type: string
          format: date
        method:
          type: ["string", "null"]
          enum: ["numerario", "transferencia", "mbway", "outro", null]
        notes:
          type: ["string", "null"]
    PaymentInput:
      type: object
      required: [quotaId, amount, date]
      properties:
        quotaId:
          type: integer
        amount:
          type: number
        date:
          type: string
          format: date
        method:
          type: ["string", "null"]
          enum: ["numerario", "transferencia", "mbway", "outro", null]
        notes:
          type: ["string", "null"]

    # ── Documents ───────────────────────────────────────────────────────────────
...